feat: implements CowApproveInfinite
