+++ packages/swapper/src/swappers/cow/CowApproveInfinite/CowApproveInfinite.ts
@@ -3,11 +3,11 @@ import { SupportedChainIds } from '@shapeshiftoss/types'
 import { ApproveInfiniteInput, SwapError, SwapErrorTypes } from '../../../api'
 import { erc20Abi } from '../../utils/abi/erc20-abi'
 import { grantAllowance } from '../../utils/helpers/helpers'
+import { CowSwapperDeps } from '../CowSwapper'
 import { MAX_ALLOWANCE } from '../utils/constants'
-import { ZrxSwapperDeps } from '../ZrxSwapper'
 
-export async function ZrxApproveInfinite(
-  { adapterManager, web3 }: ZrxSwapperDeps,
+export async function CowApproveInfinite(
+  { adapterManager, web3 }: CowSwapperDeps,
   { quote, wallet }: ApproveInfiniteInput<SupportedChainIds>
 ) {
   try {
@@ -25,7 +25,7 @@ export async function ZrxApproveInfinite(
     return allowanceGrantRequired
   } catch (e) {
     if (e instanceof SwapError) throw e
-    throw new SwapError('[ZrxApproveInfinite]', {
+    throw new SwapError('[CowApproveInfinite]', {
       cause: e,
       code: SwapErrorTypes.APPROVE_INFINITE_FAILED
     })
