+++ packages/swapper/src/swappers/cow/utils/constants.ts
@@ -1,0 +1,1 @@
+export declare const MAX_ALLOWANCE = '100000000000000000000000000'
