+++ packages/swapper/src/swappers/index.ts
@@ -1,3 +1,4 @@
 export * from './zrx/ZrxSwapper'
 export * from './thorchain/ThorchainSwapper'
 export * from './test/TestSwapper'
+export * from './cow/CowSwapper'
